Add tools and tests
Adds tools needed to construct the B-spline basis, compute angular
coefficients, perform Gaussian quadrature, and construct 1-particle +
2-particle matrix elements. Also add some simple test cases (but need to
polish them more). The cell integration algorithm for e-e interaction
has been implemented, but in its current state is does not make full use
of symmetry and sparsity, so the construction of the matrix leads to
large memory consumption (and it is quite slow).