Add support for block versions of sparse matrices
The Hamiltonian and dipole matrices are naturally assembled one symmetry
block at a time (or in parallel), so added derived types to handle this
situation. The block types can be converted to regular CS(X) matrices if
desired. Changed the sparse type used in Hamiltonian to CSR, since that
will be needed by the eigensolvers I plan to use.