Fix velocity gauge reduced dipole matrix element
The precious implementation of the reduced matrix element was incorrect.
A (hopefully) correct version has now been implemented. Added a radial
dipole type in order to still be able to use the same interface for the
velocity and length gauge setup routines. Added a procedure pointer to
select which version of the one-particle reduced matrix element should
be used.