Change to 64 bit default integer
When the basis gets too large 64 bit integers are needed for PARDISO,
and eventually the index pointers in sparse arrays. Currently this is
done with the -fdefault-integer-8 flag for gfortran. In the future it
would be better to use kind parameters for all integers, so that the
type is clear to the programmer and user. The program now requires
building stdlib and arpack with ILP64 support, and it links to the ILP64
version of MKL.