Add option for 1-e calcs and wavefunction eval
It is now possible to run calculations for one-electron systems, by
adjusting the two_el flag in the input file. A module for extracting the
spatial (partial wave resolved) wavefunction has been added. The
B-spline object created during basis setup is now stored, so that it can
be used by the remaining apps.