Add functionality for 2-e densities
Adds the option to compute the electron density from a two-electron
wavefunction on a grid in r,theta,phi. Right now it is rather slow to
evaluate, will probably need to look at ways to speed it up in the
future. Also adds the option to store vectors from the diag application,
which can be used as input for the feval program.