Add module for laser fields
Add a module for laser fields, that contains a pulse type. The module
also contains several implementations of envelopes and carriers, that
can be selected at runtime by passing the correct identifiers to the
pulse init routine. The parameters of the carriers and envelopes are
stored in two real array components of the pulse type. Also added a test
program for the fields module. Currently only the vector potential is
implemented, so propagation (also not yet implemented) would only be
possible in the velocity gauge. For length gauge propagation would
require also the derivative of the vector potential.