Add program and modules for RKH calculations
Add modules and a main program for performing Calculations in the
Rotating Kramers-Henneberger frame. The RKH_hamiltonian module computes
the matrix elements and constructs the Hamiltonian, a CAP is used to
enforce Siegert boundary conditions. Siegert states can then be found by
computing a (subset) of eigenvalues and eigenvectors. The
adiabatic_potentials module computes adiabatic potentials and angular
functions as a function of r. It also contains a subroutine to evaluate
the radial functions in the spherical harmonic basis as well as the
adiabatic basis.